Separate `concurrent_access` test case
